pyproject.toml
license = { text = "MIT" }

[tool.poetry]
name = "deep_cbc"
version = "1.0.0"
description = "Deep CBC experiment for prototype based learning benchmarks."
authors = ["Members from Hochschule Mittweida"]
readme = "README.md"
package-mode = true
packages = [
        { include = "deep_cbc" },
    ]

[tool.poetry.dependencies]
python = "^3.9"
Augmentor = "0.2.12"
jupyter = "1.0.0"
lightning = "2.2.2"
matplotlib = "3.8.4"
numpy = "1.26.4"
omegaconf = "2.3.0"
opencv-python = "^4.9.0"
pandas = "2.2.2"
pillow = "10.3.0"
pre-commit = "3.7.0"
pytest = "8.1.1"
requests = "2.31.0"
scikit-learn = "1.4.2"
scipy = "1.13.0"
seaborn = "0.13.2"
setuptools = "68.2.0"
Sphinx = "7.3.5"
tensorboard = "2.16.2"
timm = "0.4.12"
torch = "2.2.2"
torchmetrics = "1.3.2"
torchsummary = "1.5.1"
torchvision = "0.17.2"
tox = "4.14.2"
hydra-core = "1.3.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
